refactor: 게임 참여자 로직 수정 - interface -> sealed interface로 수정 - 초기 카드 배부 로직 이동
