--- src/main/kotlin/blackjack/domain/Participant.kt
@@ -1,7 +1,0 @@
-package blackjack.domain
-
-interface Participant {
-    val cards: List<TrumpCard>
-
-    fun drawCard(card: TrumpCard)
-}
