+++ src/main/kotlin/blackjack/domain/participant/Dealer.kt
@@ -1,0 +1,21 @@
+package blackjack.domain.participant
+
+import blackjack.domain.TrumpCard
+
+class Dealer : Participant {
+    private val _cards = mutableListOf<TrumpCard>()
+    override val cards: List<TrumpCard> get() = _cards.toList()
+
+    override fun getInitializedCards(
+        initializedCardCount: Int,
+        card: TrumpCard,
+    ) {
+        repeat(initializedCardCount) {
+            addCard(card)
+        }
+    }
+
+    override fun addCard(card: TrumpCard) {
+        _cards.add(card)
+    }
+}
