+++ src/main/kotlin/blackjack/domain/participant/Participant.kt
@@ -1,0 +1,14 @@
+package blackjack.domain.participant
+
+import blackjack.domain.TrumpCard
+
+sealed interface Participant {
+    val cards: List<TrumpCard>
+
+    fun getInitializedCards(
+        initializedCardCount: Int,
+        card: TrumpCard,
+    )
+
+    fun addCard(card: TrumpCard)
+}
