+++ src/test/kotlin/blackjack/PlayerTest.kt
@@ -1,9 +1,7 @@
 package blackjack
 
-import blackjack.domain.Player
-import blackjack.domain.TrumpCard
-import blackjack.domain.enums.CardTier
-import blackjack.domain.enums.Shape
+import blackjack.domain.participant.Player
+import blackjack.fixture.trumpCardFixture
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
@@ -11,8 +9,10 @@ class PlayerTest {
     @Test
     fun `플레이어가 게임을 시작하면 카드를 2장 지급받는다`() {
         val player = Player("페토")
-        player.drawCard(TrumpCard(CardTier.K, Shape.Dia))
-        player.drawCard(TrumpCard(CardTier.Ace, Shape.Heart))
-        assertThat(player.cards.size).isEqualTo(2)
+        val fixture = trumpCardFixture()
+        fixture.forEach {
+            player.addCard(it)
+        }
+        assertThat(player.cards).containsExactly(*fixture.toTypedArray())
     }
 }
